extract WorkerProfiler + remove dead helpers
- Move profiling/metrics logic out of Worker.worker.ts into WorkerProfiler.ts
- Drop unused MAX_TICKS_PER_HEARTBEAT and TransformHandler.viewOffset()
- Simplify TerritoryLayer overlay parenting; remove unused markTile() wrapper
